apps/entrance/pyproject.toml
[project]
name = "entrance"
version = "0.1.0"
description = "A local application gateway for managing and accessing self-hosted services"
authors = [
    { name = "Morten Pedersen", email = "mortenvester1@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.10.1",
    "fastapi>=0.117.1",
    "httpx>=0.27.0",
    "jinja2>=3.1.6",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.36.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=6.0.0"
]

[tool.ruff]
exclude = [".venv"]
line-length = 120

[tool.ruff.lint]
# Enable pycodestyle (E, W), pyflakes (F), and isort (I)
select = ["E", "F", "W", "I"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py files

[tool.pytest.ini_options]
addopts = "-v --tb=short"



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
